requirements.txt
flask
flask-caching
flask-session
flask-login
newrelic
peewee
pygithub
pymysql
python-dotenv
redis
requests
uwsgi
